Split into WMS, WCS and common code.  WCS exceptions supported.
