+++ datacube_wms/wsgi.py
@@ -7,5 +7,5 @@ sys.path.append("/home/phaesler/src/datacube/wms")
 # The location of the datcube config file.
 os.environ.setdefault("DATACUBE_CONFIG_PATH", "/home/phaesler/.datacube.conf.local")
 
-from datacube_wms.wms import app
+from datacube_wms.ogc import app
 application = app
